type

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -10,4 +10,4 @@ async def async_generator() -> AsyncGenerator[float, float]:
     """Asyncronous Generator function"""
     for _ in range(10):
         await asyncio.sleep(1)
-        yield random.random() * 10
+        yield float(random.random() * 10)
